This is a HACK with no actual access control value
Temporarily hides the QBO menu links from users who are not admins.
Anything with half a sitemap and a little bit of knowhow can get
around this in a heartbeat.

Need to implement proper per-project/action permissions as other
plugins do in order to convert this little fiasco to use the
scoped user attributes provided by internal access control methods.